requirements.txt
fastapi
uvicorn
pymongo
openai
pydantic
python-dotenv
requests
regex
groq